+++ packages/devtools_shared/CHANGELOG.md
@@ -1,4 +1,10 @@
-# 9.1.0-wip
+# 10.0.0-dev.0
+* **Breaking change:** rename `DevToolsExtensionConfig.path` to
+`DevToolsExtensionConfig.extensionAssetsUri` and update class documentation.
+* **Breaking change:** rename `ExtensionsApi.extensionRootPathPropertyName`
+to `ExtensionsApi.packageRootUriPropertyName`, and modify the String value
+for the parameter from 'rootPath' to 'packageRootUri'.
+
 * Return valid extensions from the `apiServeAvailableExtensions` endpoint even when
 an exception is thrown.